2nd commit: Grabbing data from Json - ok Population ListView - ok
