+++ app/src/main/res/layout/activity_the_guardian.xml
@@ -19,6 +19,10 @@
         android:layout_height="wrap_content"
         android:text="Student: Lilia Ramalho Martins" />
 
+    <Space
+        android:layout_width="match_parent"
+        android:layout_height="14dp" />
+
     <ListView
         android:id="@+id/list_results"
         android:layout_width="match_parent"
